Reduce memory usage

+++ dora/core/client/fs/src/main/java/alluxio/client/file/cache/store/LocalPageStore.java
@@ -69,19 +69,23 @@ public class LocalPageStore implements PageStore {
       boolean isTemporary) throws ResourceExhaustedException, IOException {
     Path pagePath = getPagePath(pageId, isTemporary);
     try {
-      LOG.debug("Put page: {}, page's position: {}, page's limit: {}, page's capacity: {}",
+      LOG.info("Put page: {}, page's position: {}, page's limit: {}, page's capacity: {}",
           pageId, page.position(), page.limit(), page.capacity());
       if (!Files.exists(pagePath)) {
         Path parent = Preconditions.checkNotNull(pagePath.getParent(),
             "parent of cache file should not be null");
         Files.createDirectories(parent);
-        Files.createFile(pagePath);
+        //Files.createFile(pagePath);
       }
       // extra try to ensure output stream is closed
       try (FileOutputStream fos = new FileOutputStream(pagePath.toFile(), false)) {
         fos.getChannel().write(page);
       }
-    } catch (Exception e) {
+    } catch (Throwable e) {
+      System.out.println(e.getClass().getName());
+      if (e instanceof Error) {
+        throw new IOException("Failed to write file " + pagePath + " for page " + pageId, e);
+      }
       Files.deleteIfExists(pagePath);
       if (e.getMessage() != null && e.getMessage().contains(ERROR_NO_SPACE_LEFT)) {
         throw new ResourceExhaustedException(
